refactor(api): modularize API calls into service structure
