+++ src/services/api/candidates.api.js
@@ -1,0 +1,8 @@
+import { apiClient } from "./apiClient";
+
+export const candidateService = {
+  getAll: () => apiClient('/candidates', {
+    method: 'GET',
+    headers: { 'Content-Type': 'application/json' },
+  }),
+};
