+++ src/services/api/positions.api.js
@@ -1,0 +1,13 @@
+import { apiClient } from "./apiClient";
+
+export const positionService = {
+  create: (data) => apiClient('/positions', {
+    method: 'POST',
+    headers: { 'Content-Type': 'application/json' },
+    body: JSON.stringify(data),
+  }),
+  getAll: () => apiClient('/positions', {
+    method: 'GET',
+    headers: { 'Content-Type': 'application/json' },
+  }),
+};
