+++ src/services/api/vacancies.api.js
@@ -1,0 +1,13 @@
+import { apiClient } from "./apiClient";
+
+export const vacanciesApi = {
+  getAll: () => apiClient('/vacancies', {
+    method: 'GET',
+    headers: { 'Content-Type': 'application/json' },
+  }),
+  create: (data) => apiClient('/vacancies', {
+    method: 'POST',
+    headers: { 'Content-Type': 'application/json' },
+    body: JSON.stringify(data),
+  }),
+};
